+++ src/main/java/com/mad_scientists/weird_science/block/programming_interface/base/ProgrammerBaseBlock.java
@@ -1,6 +1,5 @@
 package com.mad_scientists.weird_science.block.programming_interface.base;
 
-import com.mad_scientists.weird_science.block.tinkers_table.TinkersTableBlockEntity;
 import com.mad_scientists.weird_science.init.AllBlockEntities;
 import com.mad_scientists.weird_science.init.AllBlocks;
 import com.mad_scientists.weird_science.init.AllShapes;
@@ -40,20 +39,25 @@ public class ProgrammerBaseBlock extends BaseEntityBlock {
     public BlockEntity newBlockEntity(BlockPos pPos, BlockState pState) {
         return new ProgrammingBaseBlockEntity(pPos, pState);
     }
+    /** Places a Programming Interface above the Programming Base on placement. **/
     @Override
     public void onPlace(BlockState blockstate, Level world, BlockPos pos, BlockState oldState, boolean moving) {
         super.onPlace(blockstate, world, pos, oldState, moving);
         world.setBlock(new BlockPos(pos.getX(), pos.getY() + 1, pos.getZ()), AllBlocks.PROGRAMMING_INTERFACE.get().defaultBlockState(), 27);
     }
+    /** Breaks the Programming Interface above the Programming Base if an explosion breaks it. **/
     @Override
     public void wasExploded(Level world, BlockPos pos, Explosion e) {
         super.wasExploded(world, pos, e);
         world.destroyBlock(new BlockPos(pos.getX(), pos.getY() + 1, pos.getZ()), false);
     }
+    /** Breaks the Programming Interface above the Programming Base if a player breaks it. **/
     @Override
     public boolean onDestroyedByPlayer(BlockState blockstate, Level world, BlockPos pos, Player entity, boolean willHarvest, FluidState fluid) {
         boolean retval = super.onDestroyedByPlayer(blockstate, world, pos, entity, willHarvest, fluid);
-        world.destroyBlock(new BlockPos(pos.getX(), pos.getY() + 1, pos.getZ()), false);
+        if (world.getBlockState(new BlockPos(pos.getX(), pos.getY() + 1, pos.getZ())).getBlock() == AllBlocks.PROGRAMMING_INTERFACE.get()) {
+            world.destroyBlock(new BlockPos(pos.getX(), pos.getY() + 1, pos.getZ()), false);
+        }
         return retval;
     }
     @Override
